refactor: 로컬스토리지 로직 유틸로 이동
